+++ src/Sulu/Bundle/AdminBundle/Resources/public/js/components/list-toolbar/main.js
@@ -31,7 +31,12 @@ define([], function() {
 
         templates = {
             default: function() {
-                return [
+                return this.sandbox.sulu.buttons.get(
+                    {'settings': {
+                        dropdownItems: [{type: 'columnOptions'}]
+                    }}
+                );
+                /*return [
                     {
                         id: 'settings',
                         icon: 'gear',
@@ -42,20 +47,17 @@ define([], function() {
                             }
                         ]
                     }
-                ];
+                ];*/
             },
             defaultEditable: function() {
-                return templates.default.call(this).concat([
-                    {
-                        id: 'edit',
-                        icon: 'pencil',
-                        position: 25,
-                        disabled: true,
+                return templates.default.call(this).concat(this.sandbox.sulu.buttons.get(
+                    {'edit': {
                         callback: function() {
                             this.sandbox.emit('sulu.list-toolbar.edit');
                         }.bind(this)
-                    }
-                ]);
+                    }}
+                ));
+
             },
             defaultNoSettings: function() {
                 var defaults = templates.default.call(this);
@@ -68,50 +70,7 @@ define([], function() {
                 return defaults;
             },
             changeable: function() {
-                return [
-                    {
-                        id: 'change',
-                        icon: 'th-large',
-                        dropdownOptions: {
-                            markSelected: true
-                        },
-                        dropdownItems: [
-                            {
-                                id: 'small-thumbnails',
-                                title: this.sandbox.translate('sulu.list-toolbar.small-thumbnails'),
-                                callback: function() {
-                                    this.sandbox.emit('sulu.list-toolbar.change.thumbnail-small');
-                                }.bind(this)
-                            },
-                            {
-                                id: 'big-thumbnails',
-                                title: this.sandbox.translate('sulu.list-toolbar.big-thumbnails'),
-                                callback: function() {
-                                    this.sandbox.emit('sulu.list-toolbar.change.thumbnail-large');
-                                }.bind(this)
-                            },
-                            {
-                                id: 'table',
-                                title: this.sandbox.translate('sulu.list-toolbar.table'),
-                                callback: function() {
-                                    this.sandbox.emit('sulu.list-toolbar.change.table');
-                                }.bind(this)
-                            }
-                        ]
-                    }
-                ];
-            },
-            defaultEditableList: function() {
-                var defaults = templates.default.call(this);
-                defaults.splice(1, 0, {
-                    icon: 'floppy-o',
-                    disabled: true,
-                    id: 'save',
-                    callback: function() {
-                        this.sandbox.emit('sulu.list-toolbar.save');
-                    }.bind(this)
-                });
-                return defaults;
+                return this.sandbox.sulu.buttons.get('layout');
             }
         },
         listener = {
@@ -131,14 +90,6 @@ define([], function() {
                 this.sandbox.on('sulu.list-toolbar.' + instanceName + 'edit.state-change', function(enable) {
                     postfix = !!enable ? 'enable' : 'disable';
                     this.sandbox.emit('husky.toolbar.' + instanceName + 'item.' + postfix, 'edit', false);
-                }.bind(this));
-            },
-            defaultEditableList: function() {
-                var instanceName = this.options.instanceName ? this.options.instanceName + '.' : '';
-                listener.default.call(this);
-
-                this.sandbox.on('husky.datagrid.data.changed', function() {
-                    this.sandbox.emit('husky.toolbar.' + instanceName + 'item.enable', 'save');
                 }.bind(this));
             }
         },